Reorganized the pytest test infrastructure to make it more friendly to different FusionDefintion function patterns.
